Support Prometheus-style labels, arbitrary histograms.
This is a pretty major overhaul of the public API.

The `Tacho` type holds a metrics pipeline:

`Metrics` provides a closeable, scopable handle for applications to produce
metrics. It provides a `Scope`, which creates labeled `MetricKey`s. These keys
identify a metric and its type, but accept no values.  Keys are to be
initialized early in program execution, and may be cloned cheaply (without
copying the underling labels). `Metrics` also produces RAII-guarded `Recorder`
instances. Values are passed to the the recorder with a key. When the recorder
instance is dropped, a `Sample` is sent on the recorder's mpsc channel.

An `Aggregator` processes a stream of `Sample`s cooperatively (i.e. yielding
back to the reactor after processing a batch of samples). Finally, the
aggregator publishes a `Report`.  The consumer of this `Report` may reset it as
it sees fit.