Adopt a synchronous API
Previously, our metrics pipeline was a bit complex and CPU-intensive, relying
on a thread-safe channel and aggregator to shuttle metrics into a reporting
thread.

This can be simplified by removing the channel in favor of an `Arc<RwLock<_>>`.
While this may cause contention for many-threaded applications, this is fine
for our primary use case where a single thread is responsible for producing all
metrics.  The Metrics/Recorder/Aggregator/Report API has been simplified to a
simpler Scope/Reporter API.

`Scope` is responsible for creating writeable metrics with contextual labels.

`Reporter` supports two means of obtaining a `Report` on the current state of
metrics:
- Reporter::peek -- borrows a read-only view of the underlying registry,
  intended for short-lived access
- Reporter::take -- which copies counters and flushes gauges and stats into a
  new Report that may be held indefinitely

Tests have been added to validate both report_peek and report_reset.